+++ lib-test/api/lib-test.api
@@ -13,6 +13,14 @@ public final class app/cash/nostrino/crypto/ArbKeys {
 	public final fun getArbSecKey ()Lio/kotest/property/Arb;
 }
 
+public final class app/cash/nostrino/model/ArbEventContent {
+	public static final field INSTANCE Lapp/cash/nostrino/model/ArbEventContent;
+	public final fun getArbReaction ()Lio/kotest/property/Arb;
+	public final fun getArbTextNote ()Lio/kotest/property/Arb;
+	public final fun getArbUserMetaData ()Lio/kotest/property/Arb;
+	public final fun getArbZapRequest ()Lio/kotest/property/Arb;
+}
+
 public final class app/cash/nostrino/model/ArbTags {
 	public static final field INSTANCE Lapp/cash/nostrino/model/ArbTags;
 	public final fun getArbAmountTag ()Lio/kotest/property/Arb;
