+++ lib-test/src/main/resources/emojis.txt
