+++ lib/src/test/kotlin/app/cash/nostrino/model/ReactionTest.kt
@@ -1,17 +1,12 @@
 package app.cash.nostrino.model
 
-import app.cash.nostrino.ArbPrimitive.arbByteString32
-import app.cash.nostrino.crypto.PubKeyTest.Companion.arbPubKey
 import app.cash.nostrino.crypto.SecKeyTest
+import app.cash.nostrino.model.ArbEventContent.arbReaction
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.collections.shouldContainExactly
 import io.kotest.matchers.shouldBe
 import io.kotest.property.Arb
-import io.kotest.property.arbitrary.choose
-import io.kotest.property.arbitrary.element
-import io.kotest.property.arbitrary.flatMap
 import io.kotest.property.arbitrary.pair
-import io.kotest.property.arbitrary.triple
 import io.kotest.property.checkAll
 
 class ReactionTest : StringSpec({
@@ -27,19 +22,6 @@ class ReactionTest : StringSpec({
   }
 }) {
   companion object {
-    private val emojis: List<Int> by lazy {
-      Companion::class.java.getResource("/emojis.txt").readText()
-        .lines().filterNot { it.isEmpty() }.map { it.codePointAt(0) }
-    }
-    private val arbEmoji = Arb.element(emojis)
-    val arbReaction = Arb.triple(arbByteString32, arbPubKey, arbEmoji)
-      .flatMap { (e, p, c) ->
-        Arb.choose(
-          3 to Upvote(e, p),
-          3 to Downvote(e, p),
-          1 to EmojiReact(String(Character.toChars(c)), e, p)
-        )
-      }
     private val testData = Arb.pair(arbReaction, SecKeyTest.arbSecKey)
   }
 }
